requirements.txt
zip-files
npyscreen
python-telegram-bot
pycryptodome